fix: Address multiple defects and code quality issues
This commit fixes several defects and code quality issues that were identified during a defect analysis of the codebase.

The following changes have been made:

- **Floating Point Precision:** Refactored `UbkCalculateGMDServiceImp` to use `BigDecimal` for financial calculations, preventing rounding errors.
- **NullPointerException:** Added a null check in `PersonServiceImp` to prevent a `NullPointerException` when handling a null nationality from the passport service.
- **Incorrect Null/Empty Check:** Corrected the null/empty check for the death date in `PersonServiceImp`.
- **Duplicated Code:** Refactored `PersonServiceImp` to remove duplicated code for finding an existing person.
- **Hardcoded IP Addresses:** Replaced hardcoded localhost IP addresses in `RequestUtils` with constants.
- **Unnecessary Imports:** Removed several unused imports from various classes.
- **Logging Issue:** Corrected a logging message in `UbkApplicationServiceImpl`.

+++ ubk/src/main/java/kg/mlsp/ubk/service/UbkCalculateGMDService.java
@@ -1,5 +1,8 @@
 package kg.mlsp.ubk.service;
+
+import java.math.BigDecimal;
+
 public interface UbkCalculateGMDService {
-    Double calculateGMD(Integer applicationId);
+    BigDecimal calculateGMD(Integer applicationId);
 
 }
